restructuring with lastest cosmos-sdk
