+++ go.mod
@@ -1,15 +1,15 @@
-go 1.14
+go 1.15
 
 module github.com/tendermint/liquidity
 
 require (
-	github.com/cosmos/cosmos-sdk v0.34.4-0.20200914022129-c26ef79ed0a2
+	github.com/cosmos/cosmos-sdk v0.34.4-0.20201006170641-3e6089dc0ea6
 	github.com/gogo/protobuf v1.3.1
 	github.com/golang/protobuf v1.4.2
 	github.com/gorilla/mux v1.8.0
-	github.com/grpc-ecosystem/grpc-gateway v1.14.8
+	github.com/grpc-ecosystem/grpc-gateway v1.15.0
 	github.com/spf13/cobra v1.0.0
-	github.com/tendermint/tendermint v0.34.0-rc3.0.20200907055413-3359e0bf2f84
+	github.com/tendermint/tendermint v0.34.0-rc4
 	google.golang.org/genproto v0.0.0-20200825200019-8632dd797987
 	google.golang.org/grpc v1.32.0
 	gopkg.in/yaml.v2 v2.3.0
