+++ x/liquidity/spec/02_state.md
